requirements.txt
torch>=2.0.0
transformers>=4.30.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
jieba>=0.42.1
synonyms>=3.16.0
textattack>=0.3.8
tqdm>=4.65.0
pyyaml>=6.0
scipy>=1.10.0
nltk>=3.8.0